% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chebycenter.R
\name{chebycenter}
\alias{chebycenter}
\title{chebycenter}
\usage{
chebycenter(A, b)
}
\arguments{
\item{A}{a matrix}

\item{b}{a vector of length equals to nrow(A)}
}
\value{
a vector corresponding to the centroid of the polytope
}
\description{
This code is a translation of the matlab code developped by Tim Bernham - University of Queensland
(https://ch.mathworks.com/matlabcentral/fileexchange/34208-uniform-distribution-over-a-convex-polytope
It computes the centroid of the complex polytope defined by \eqn{A \cdot x \leqslant   b}
}
\examples{
n <- 20
A1 <- -diag(n)
b1 <- as.matrix(rep(0,n))
A2 <- diag(n)
b2 <- as.matrix(rep(1,n))
A <- rbind(A1,A2)
b <- rbind(b1,b2)
X0 <- chebycenter(A,b)
}
