% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libPath.R
\name{libPath}
\alias{libPath}
\title{libPath
This function provides the path of the compiled library}
\usage{
libPath()
}
\value{
path of the compiled library
}
\description{
libPath
This function provides the path of the compiled library
}
\details{
This function provides the path of the compiled library so that the c++ code can be directly called from c++ routines in other other packages. You need to specify
something like PKG_LIBS += '${R_HOME}/bin/Rscript -e "cat(cpgsR::libPath())"'
}
\examples{
libPath()

}
